Move IngestID into a different namespace
