+++ bag_unpacker/src/main/scala/uk/ac/wellcome/platform/archive/bagunpacker/builders/BagLocationBuilder.scala
@@ -3,7 +3,7 @@ package uk.ac.wellcome.platform.archive.bagunpacker.builders
 import java.nio.file.Paths
 
 import uk.ac.wellcome.platform.archive.bagunpacker.config.models.BagUnpackerWorkerConfig
-import uk.ac.wellcome.platform.archive.common.IngestID
+import uk.ac.wellcome.platform.archive.common.ingests.models.IngestID
 import uk.ac.wellcome.platform.archive.common.storage.models.StorageSpace
 import uk.ac.wellcome.storage.ObjectLocation
