+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/PipelinePayload.scala
@@ -3,7 +3,7 @@ package uk.ac.wellcome.platform.archive.common
 import java.time.Instant
 
 import uk.ac.wellcome.platform.archive.common.bagit.models.ExternalIdentifier
-import uk.ac.wellcome.platform.archive.common.ingests.models.Ingest
+import uk.ac.wellcome.platform.archive.common.ingests.models.{Ingest, IngestID}
 import uk.ac.wellcome.platform.archive.common.storage.models.StorageSpace
 import uk.ac.wellcome.storage.ObjectLocation
