+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/ingests/models/CallbackNotification.scala
@@ -2,8 +2,6 @@ package uk.ac.wellcome.platform.archive.common.ingests.models
 
 import java.net.URI
 
-import uk.ac.wellcome.platform.archive.common.IngestID
-
 case class CallbackNotification(
   ingestId: IngestID,
   callbackUri: URI,
