+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/ingests/models/IngestID.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.common
+package uk.ac.wellcome.platform.archive.common.ingests.models
 
 import java.util.UUID
