+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/ingests/models/IngestUpdate.scala
@@ -1,6 +1,5 @@
 package uk.ac.wellcome.platform.archive.common.ingests.models
 
-import uk.ac.wellcome.platform.archive.common.IngestID
 import uk.ac.wellcome.platform.archive.common.bagit.models.BagId
 import uk.ac.wellcome.platform.archive.common.bagit.models.error.ArchiveError
