+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/versioning/DynamoIngestVersionManagerDao.scala
@@ -3,9 +3,9 @@ package uk.ac.wellcome.platform.archive.common.versioning
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB
 import com.gu.scanamo.syntax._
 import com.gu.scanamo.{Scanamo, Table}
-import uk.ac.wellcome.platform.archive.common.IngestID
 import uk.ac.wellcome.platform.archive.common.bagit.models.ExternalIdentifier
 import uk.ac.wellcome.platform.archive.common.bagit.models.ExternalIdentifier._
+import uk.ac.wellcome.platform.archive.common.ingests.models.IngestID
 import uk.ac.wellcome.storage.dynamo._
 
 import scala.util.Try
