+++ common/src/main/scala/uk/ac/wellcome/platform/archive/common/versioning/IngestVersionManagerDao.scala
@@ -1,7 +1,7 @@
 package uk.ac.wellcome.platform.archive.common.versioning
 
-import uk.ac.wellcome.platform.archive.common.IngestID
 import uk.ac.wellcome.platform.archive.common.bagit.models.ExternalIdentifier
+import uk.ac.wellcome.platform.archive.common.ingests.models.IngestID
 
 import scala.util.Try
