+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/fixtures/RandomThings.scala
@@ -5,8 +5,8 @@ import java.nio.file.Paths
 import java.time.LocalDate
 import java.util.UUID
 
-import uk.ac.wellcome.platform.archive.common.IngestID
 import uk.ac.wellcome.platform.archive.common.bagit.models._
+import uk.ac.wellcome.platform.archive.common.ingests.models.IngestID
 import uk.ac.wellcome.storage.fixtures.S3.Bucket
 
 import scala.util.Random
