+++ common/src/test/scala/uk/ac/wellcome/platform/archive/common/ingests/fixtures/IngestUpdateAssertions.scala
@@ -4,7 +4,6 @@ import grizzled.slf4j.Logging
 import org.scalatest.{Assertion, Inside, Matchers}
 import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.memory.MemoryMessageSender
-import uk.ac.wellcome.platform.archive.common.IngestID
 import uk.ac.wellcome.platform.archive.common.bagit.models.BagId
 import uk.ac.wellcome.platform.archive.common.ingests.models._
