+++ display/src/main/scala/uk/ac/wellcome/platform/archive/display/DisplayIngest.scala
@@ -4,7 +4,6 @@ import java.net.{URI, URL}
 import java.util.UUID
 
 import io.circe.generic.extras.JsonKey
-import uk.ac.wellcome.platform.archive.common.IngestID
 import uk.ac.wellcome.platform.archive.common.ingests.models._
 
 sealed trait DisplayIngest
