+++ ingests_common/src/test/scala/uk/ac/wellcome/platform/archive/common/ingests/IngestTrackerTest.scala
@@ -13,7 +13,6 @@ import org.mockito.Mockito.when
 import org.scalatest.{FunSpec, TryValues}
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.mockito.MockitoSugar
-import uk.ac.wellcome.platform.archive.common.IngestID
 import uk.ac.wellcome.platform.archive.common.bagit.models.BagId
 import uk.ac.wellcome.platform.archive.common.generators.IngestGenerators
 import uk.ac.wellcome.platform.archive.common.ingests.fixtures.IngestTrackerFixture
