+++ notifier/src/test/scala/uk/ac/wellcome/platform/archive/notifier/services/PrepareNotificationServiceTest.scala
@@ -4,10 +4,9 @@ import akka.http.scaladsl.model.HttpResponse
 import org.scalatest.prop.TableDrivenPropertyChecks._
 import org.scalatest.{Assertion, FunSpec, Matchers}
 import uk.ac.wellcome.akka.fixtures.Akka
-import uk.ac.wellcome.platform.archive.common.IngestID
 import uk.ac.wellcome.platform.archive.common.fixtures.RandomThings
 import uk.ac.wellcome.platform.archive.common.ingest.fixtures.TimeTestFixture
-import uk.ac.wellcome.platform.archive.common.ingests.models.Callback
+import uk.ac.wellcome.platform.archive.common.ingests.models.{Callback, IngestID}
 
 import scala.util.{Failure, Success, Try}
